+++ layers/WPSchema/packages/customposts/src/Hooks/QueryHookSet.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PoPWPSchema\CustomPosts\Hooks;
 
 use PoP\Hooks\AbstractHookSet;
-use PoPSchema\CustomPostsWP\TypeAPIs\CustomPostTypeAPI;
+use PoPSchema\CustomPostsWP\TypeAPIs\AbstractCustomPostTypeAPI;
 use PoPWPSchema\CustomPosts\Constants\CustomPostOrderBy;
 
 class QueryHookSet extends AbstractHookSet
@@ -13,9 +13,25 @@ class QueryHookSet extends AbstractHookSet
     protected function init(): void
     {
         $this->getHooksAPI()->addFilter(
-            CustomPostTypeAPI::HOOK_ORDERBY_QUERY_ARG_VALUE,
+            AbstractCustomPostTypeAPI::HOOK_ORDERBY_QUERY_ARG_VALUE,
             [$this, 'getOrderByQueryArgValue']
         );
+        $this->getHooksAPI()->addFilter(
+            AbstractCustomPostTypeAPI::HOOK_QUERY,
+            [$this, 'convertCustomPostsQuery'],
+            10,
+            2
+        );
+    }
+
+    public function convertCustomPostsQuery($query, array $options): array
+    {
+        // `null` is an accepted value
+        if (array_key_exists('has-password', $query)) {
+            $query['has_password'] = $query['has-password'];
+            unset($query['has-password']);
+        }
+        return $query;
     }
 
     public function getOrderByQueryArgValue(string $orderBy): string